+++ CustomerPatronage/Training/HistoricalPurchaseData.cs
@@ -5,6 +5,6 @@ namespace CustomerPatronage.Training
         public string CustomerId { get; set; } = string.Empty;
         public string CustomerName { get; set; } = string.Empty;
         public DateTime PurchaseDate { get; set; }
-        public decimal Spend { get; set; }
+        public float Spend { get; set; }
     }
 }
