+++ CustomerPatronage/Training/ModelMetadata.cs
@@ -2,7 +2,7 @@ namespace CustomerPatronage.Training
 {
     public class ModelMetadata
     {
-        public decimal AverageSpend { get; set; }
+        public float AverageSpend { get; set; }
 
         public float AverageFrequency { get; set; }
     }
